+++ packages/react-bindings/src/hooks/index.ts
@@ -2,7 +2,6 @@ import * as CallStateHooks from './callStateHooks';
 
 export * from './permissions';
 export * from './store';
-export * from './callStateHooks';
 
 /**
  * A hook-alike function that exposes all call state hooks.