+++ packages/react-native-sdk/ios/StreamVideoReactNative.m
@@ -1,0 +1,19 @@
+#import <React/RCTBridgeModule.h>
+#import "StreamVideoReactNative.h"
+#import "WebRTCModule.h"
+#import "WebRTCModuleOptions.h"
+
+@implementation StreamVideoReactNative
+RCT_EXPORT_MODULE();
+
++(BOOL)requiresMainQueueSetup {
+    return NO;
+}
+
++(void)setup {
+    RTCDefaultVideoEncoderFactory *videoEncoderFactory = [[RTCDefaultVideoEncoderFactory alloc] init];
+    RTCVideoEncoderFactorySimulcast *simulcastVideoEncoderFactory = [[RTCVideoEncoderFactorySimulcast alloc] initWithPrimary:videoEncoderFactory fallback:videoEncoderFactory];
+    WebRTCModuleOptions *options = [WebRTCModuleOptions sharedInstance];
+    options.videoEncoderFactory = simulcastVideoEncoderFactory;
+}
+@end