+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,42 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.2.3](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.2.2...@stream-io/video-react-bindings-0.2.3) (2023-08-16)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.1.0`
+
+### Features
+
+* use new device API in RN SDK and move to @stream-io/react-native-webrtc ([#925](https://github.com/GetStream/stream-video-js/issues/925)) ([8442d82](https://github.com/GetStream/stream-video-js/commit/8442d821a8eb97cb4be6e6d71b64337c04a86a15))
+
+### [0.2.2](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.2.1...@stream-io/video-react-bindings-0.2.2) (2023-08-16)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.1.0`
+### [0.2.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.2.0...@stream-io/video-react-bindings-0.2.1) (2023-08-16)
+
+
+### Bug Fixes
+
+* correctly report `live` state of the call ([#940](https://github.com/GetStream/stream-video-js/issues/940)) ([75f36e1](https://github.com/GetStream/stream-video-js/commit/75f36e1e955da01b683a299a921b8abc74e87982)), closes [#931](https://github.com/GetStream/stream-video-js/issues/931)
+
+## [0.2.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.1.18...@stream-io/video-react-bindings-0.2.0) (2023-08-16)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.1.0`
+
+### ⚠ BREAKING CHANGES
+
+* Call State reorganization (#931)
+
+### Features
+
+* Call State reorganization ([#931](https://github.com/GetStream/stream-video-js/issues/931)) ([441dbd4](https://github.com/GetStream/stream-video-js/commit/441dbd4ffb8c851abb0ca719be143a1e80d1418c)), closes [#917](https://github.com/GetStream/stream-video-js/issues/917)
+
 ### [0.1.18](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.1.17...@stream-io/video-react-bindings-0.1.18) (2023-08-14)
 
 ### Dependency Updates